+++ app/models/local_authority.rb
@@ -20,6 +20,7 @@ class LocalAuthority < ApplicationRecord
     has_many :policy_references
     has_many :requirements
     has_many :api_users
+    has_many :application_types
   end
 
   with_options through: :planning_applications do
