+++ engines/bops_api/app/services/bops_api/application/parsers/base_parser.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module BopsApi
+  module Application
+    module Parsers
+      class BaseParser
+        attr_reader :params, :local_authority
+
+        def initialize(params, local_authority:)
+          @params = params
+          @local_authority = local_authority
+        end
+      end
+    end
+  end
+end
