+++ engines/bops_api/app/services/bops_api/application/parsers/proposal_parser.rb
@@ -3,13 +3,7 @@
 module BopsApi
   module Application
     module Parsers
-      class ProposalParser
-        attr_reader :params
-
-        def initialize(params)
-          @params = params
-        end
-
+      class ProposalParser < BaseParser
         def parse
           {
             description: params[:description],
