+++ engines/bops_api/spec/services/application/parsers/pre_assessment_parser_spec.rb
@@ -4,8 +4,10 @@ require "rails_helper"
 
 RSpec.describe BopsApi::Application::Parsers::PreAssessmentParser do
   describe "#parse" do
+    let(:local_authority) { create(:local_authority, :default) }
+
     let(:parse_pre_assessment) do
-      described_class.new(params).parse
+      described_class.new(params, local_authority:).parse
     end
 
     context "with valid params" do
