+++ spec/system/planning_applications/consulting/send_letters_to_neighbours_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe "Send letters to neighbours", type: :system, js: true do
   let(:api_user) { create(:api_user, name: "PlanX") }
   let(:default_local_authority) { create(:local_authority, :default) }
   let(:assessor) { create(:user, :assessor, local_authority: default_local_authority) }
-  let!(:application_type) { create(:application_type, :prior_approval) }
+  let!(:application_type) { create(:application_type, :prior_approval, local_authority: default_local_authority) }
 
   let(:planning_application) do
     create(:planning_application,
