+++ spec/system/planning_applications/sorting_spec.rb
@@ -4,8 +4,8 @@ require "rails_helper"
 
 RSpec.describe "Planning application sorting" do
   let!(:default_local_authority) { create(:local_authority, :default) }
-  let!(:application_type_ldc_proposed) { create(:application_type, :ldc_proposed) }
-  let!(:application_type_prior_approval) { create(:application_type, :prior_approval) }
+  let!(:application_type_ldc_proposed) { create(:application_type, :ldc_proposed, local_authority: default_local_authority) }
+  let!(:application_type_prior_approval) { create(:application_type, :prior_approval, local_authority: default_local_authority) }
   let(:assessor) { create(:user, :assessor, local_authority: default_local_authority) }
 
   before do
